2007-06-15  Martin Hunt  <hunt@redhat.com> 	From Quentin Barnes. 	* loc2c-runtime.h: Latest arm marcos. 	* stack-arm.c (__stp_stack_print): Add a cast. 	* regs.c (_stp_ret_addr): Define for arm.

+++ runtime/stack-arm.c
@@ -59,7 +59,7 @@ static void __stp_stack_print (struct pt_regs *regs, int verbose, int levels)
 			_stp_symbol_print((unsigned long)pc);
 			_stp_print_char('\n');
 		} else {
-			_stp_printf("%08lx ", pc);
+			_stp_printf("%08lx ", (unsigned long)pc);
 		}
 
 		/* Sanity check the next_fp. */
